Improved i18n type safety
