+++ src/@types/react-i18next.d.ts
@@ -1,18 +1,6 @@
 import "react-i18next"
 
-import en from "@/src/locales/en.json"
-
-export type Translation = typeof en
-
-type Paths<T, K extends keyof T = keyof T> = K extends string | number
-  ? T[K] extends Record<string, unknown>
-    ? `${K}.${Paths<T[K]>}`
-    : K
-  : never
-
-export type TranslationKey = Paths<Translation>
-
-export type ValidateTranslationKeys<T extends TranslationKey> = T
+import { Translation } from "@/src/shared/types/i18n"
 
 declare module "i18next" {
   interface CustomTypeOptions {
